mac/agent_charts_screen/requirements.txt
mss>=9.0
Pillow>=10.0
httpx>=0.27
python-dotenv>=1.0
pyobjc-framework-Quartz>=10.1
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pyautogui>=0.9.54
